Add a releases.csv table for overall table metadata
This table helps store misc useful metadata about a release in a table,
so it will be easier to use when say importing to a spreadsheet program.

Python requirements is now mandated at 3.11+ because it has a better
datetime.fromisoformat() util, and I don't want to add dateutil which is
an external pip dependency.